--- reduct/enf/.gitignore
@@ -1,2 +1,0 @@
-__pycache__
-.vscode
