--- reduct/enf/Tests/__init__.py
@@ -1,4 +1,0 @@
-from Tests.TransformationTestCases import *
-from Tests.GeneralTestCases import *
-from Tests.TestHelpers import *
-from Tests.EnfRuleCheckers import *
